requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
twilio==6.56.0
urllib3==1.26.4
Werkzeug==1.0.1
